+++ src/chinodeco/debug/__init__.py
@@ -1,6 +1,8 @@
 # !/usr/bin/env Python3
 # -*- coding:utf-8 -*-
 
+__all__ = ["debug", "trycatch", "DEBUG", "UnknownCommandError", "UnknownParameterError", "ArgumentCountError"]
+
 from .debugger import (
     DEBUG,
     debug,