+++ src/chinodeco/debug/errors.py
@@ -1,6 +1,8 @@
 # !/usr/bin/env Python3
 # -*- coding:utf-8 -*-
 
+MODULE = "chinodeco.debug.errors"
+
 class UnknownCommandError(Exception):
     pass