+++ src/chinodeco/decodsl/__init__.py
@@ -1,6 +1,8 @@
 # !/usr/bin/env Python3
 # -*- coding:utf-8 -*-
 
+__all__ = ["when", "whileloop", "foreach", "CommandDispatcher"]
+
 from .control import (
     when,
     whileloop,