+++ src/chinodeco/parameter/__init__.py
@@ -1,7 +1,12 @@
 # !/usr/bin/env Python3
 # -*- coding:utf-8 -*-
 
-from .adders import (
+__all__ = ["setargs", "addprefix", "addsuffix", "mapargs", "filterargs"]
+
+from .pretreat import (
+    setargs,
     addprefix,
-    addsuffix
+    addsuffix,
+    mapargs,
+    filterargs
 )